pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "msb-extractor"
version = "0.1.0"
description = "Extract your MyStrengthBook training data into a clean, portable spreadsheet."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "Matthias Bernhard" }]
keywords = ["mystrengthbook", "powerlifting", "training-log", "scraper", "xlsx"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
]
dependencies = [
    "beautifulsoup4>=4.12",
    "lxml>=5.1",
    "pydantic>=2.6",
    "openpyxl>=3.1",
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
    "types-beautifulsoup4",
]

[project.scripts]
msb-extractor = "msb_extractor.cli:app"

[project.urls]
Homepage = "https://github.com/matthiasbernhard/msb-extractor"
Issues = "https://github.com/matthiasbernhard/msb-extractor/issues"
Source = "https://github.com/matthiasbernhard/msb-extractor"

[tool.hatch.build.targets.wheel]
packages = ["src/msb_extractor"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/scraper",
    "/docs",
    "/README.md",
    "/LICENSE",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=msb_extractor",
    "--cov-report=term-missing",
]

[tool.coverage.run]
source = ["src/msb_extractor"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "E501",   # line-too-long (black-style wrapping is enough)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
disallow_untyped_defs = true
files = ["src/msb_extractor"]

[[tool.mypy.overrides]]
module = ["openpyxl.*"]
ignore_missing_imports = true
